장비 수정
장비 착용하면 기본 스탯이 적용이 안되던 버그 수정

장비 희귀도가 5:3:2로 나오게 끔 수정

장비 옵션 수치 조정

+++ Assets/3.Script/Equipment/EquipmentStst.cs
@@ -30,12 +30,12 @@ public class EquipmentStst
     public float CriticalChance { get; set; }
     public float CriticalDamage { get; set; }
 
-    public int MinAttackDamage { get; set; }
-    public int MaxAttackDamage { get; set; }
+    public float MinAttackDamage { get; set; }
+    public float MaxAttackDamage { get; set; }
     public float IncreaseAttackDamage { get; set; }
 
-    public int MinSpellDamage { get; set; }
-    public int MaxSpellDamage { get; set; }
+    public float MinSpellDamage { get; set; }
+    public float MaxSpellDamage { get; set; }
     public float IncreaseSpellDamage { get; set; }
 
     public void Initialize()
